🤖 refactor(site/src/pages): nest agent settings pages
