+++ site/src/pages/AgentsPage/AgentSettingsPage/AgentSettingsAPIKeysPage/AgentSettingsAPIKeysPageView.tsx
@@ -8,7 +8,7 @@ import { ConfirmDialog } from "#/components/Dialog/ConfirmDialog/ConfirmDialog";
 import { EmptyState } from "#/components/EmptyState/EmptyState";
 import { Input } from "#/components/Input/Input";
 import { Loader } from "#/components/Loader/Loader";
-import { SectionHeader } from "./components/SectionHeader";
+import { SectionHeader } from "../../components/SectionHeader";
 
 const API_KEY_PLACEHOLDER = "••••••••••••••••";
