+++ site/src/pages/AgentsPage/AgentSettingsPage/AgentSettingsGeneralPage/AgentSettingsGeneralPageView.tsx
@@ -1,16 +1,16 @@
 import type { FC } from "react";
 import type { UseMutateFunction } from "react-query";
 import type * as TypesGen from "#/api/typesGenerated";
-import { ChatFullWidthSettings } from "./components/ChatFullWidthSettings";
-import { ChatSendShortcutSettings } from "./components/ChatSendShortcutSettings";
+import { ChatFullWidthSettings } from "../../components/ChatFullWidthSettings";
+import { ChatSendShortcutSettings } from "../../components/ChatSendShortcutSettings";
 import {
 	CodeDiffDisplaySettings,
 	ShellToolDisplaySettings,
 	ThinkingDisplaySettings,
-} from "./components/DisplayModeSettings";
-import { PersonalInstructionsSettings } from "./components/PersonalInstructionsSettings";
-import { SectionHeader } from "./components/SectionHeader";
-import { UserChatDebugLoggingSettings } from "./components/UserChatDebugLoggingSettings";
+} from "../../components/DisplayModeSettings";
+import { PersonalInstructionsSettings } from "../../components/PersonalInstructionsSettings";
+import { SectionHeader } from "../../components/SectionHeader";
+import { UserChatDebugLoggingSettings } from "../../components/UserChatDebugLoggingSettings";
 
 export interface AgentSettingsGeneralPageViewProps {
 	userPromptData: TypesGen.UserChatCustomPrompt | undefined;
