+++ site/src/pages/AgentsPage/AgentSettingsPage/AgentSettingsLayout.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
 import { Outlet, useLocation } from "react-router";
 import { ScrollArea } from "#/components/ScrollArea/ScrollArea";
-import { AgentPageHeader } from "./components/AgentPageHeader";
+import { AgentPageHeader } from "../components/AgentPageHeader";
 
 const AgentSettingsLayout: FC = () => {
 	const location = useLocation();
